fix: checkComCheckAppMiddleware moved  in middleware
