+++ middleware.js
@@ -1,5 +1,8 @@
 import { createMiddlewareClient } from '@supabase/auth-helpers-nextjs'
 import { NextResponse } from 'next/server'
+import { checkTokenExistsInDatabase } from './checkToken'
+
+const COM_CHECK_APP = process.env.COM_CHECK_APP
 
 export async function middleware(req) {
   // We need to create a response and hand it to the supabase client to be able to modify the response headers.
@@ -26,4 +29,21 @@ export async function middleware(req) {
 
 export const config = {
   matcher: '/projects/:path*',
+}
+
+export const checkComCheckAppMiddleware = async (req, res, next) => {
+  try {
+    if (!COM_CHECK_APP) {
+      const tokenResult = await checkTokenExistsInDatabase(req)
+      console.log('API GOOD')
+
+      if (!tokenResult) {
+        return res.status(401).json({ error: 'Unauthorized' })
+      }
+    }
+    console.log('GOOD')
+    next()
+  } catch (error) {
+    return res.status(500).json({ error: 'Internal Server Error' })
+  }
 }
